Added Google TTS, removed some redundant functions, added call to action options, corrected issue with file sorting when combining files

+++ askreddit/main.py
@@ -1,30 +1,44 @@
 from types import NoneType
 from utils.redditScrape import scrapeComments
-from utils.audioGenerator import soundifyAuthor, soundifyComment, soundifyFollow, soundifyMakeComment
+# from utils.audioGenerator import soundifyAuthor, soundifyComment
 # from utils.audioEleven import soundifyAuthor, soundifyComment
+from utils.audioGTTS import soundifyAuthor, soundifyComment
 from utils.captionCreate import commentImage, titleImage, commentBlankImage
 from utils.videoCreate import createVideo
 import shutil
 import os
 import time
 
+# input of how many posts to scrape
+num_of_posts = int(input('How many videos would you like to create?\n'))
+
 # Gets posts and top x comments from selected subreddit
 ### subreddit, number of posts, timeframe
-
-
-for i in range(5):
+for i in range(num_of_posts):
     #for file in os.listdir("temp"):
     #    os.remove("temp/"+file)
-    
+
     postnum = i + 1
     # scrape best post i to create movie
     post = scrapeComments("askreddit", postnum, "day")
-    # debug line
-    print(post[0].author)
+    # DEBUG
+    # print(post[0].author)
 
     # get post author
     asker = str(post[0].author)
 
+    skip = False
+
+    # skip if video already exists for asker - probably a cleaner way to do this
+    for files in os.listdir('../exports'):
+        if asker + '.mp4' in files:
+            print ('File already exists for thread.\n')
+            skip = True
+    if skip:
+        print ('Skipping post by: ' + asker)
+        continue
+
+    # delete post directory if already exists
     if os.path.isdir(asker):
         shutil.rmtree(asker)
     
@@ -34,44 +48,49 @@ for i in range(5):
     # create directory to store files
     os.makedirs(asker)
 
-    # if len(post) > 1:
-    #     post.insert(1, 'Don\'t forget to follow!')
-    #     post[1].author = asker
-    #     commentBlankImage(post[1], 1, asker)
+    if len(post) > 1:
+        post.insert(1, {'author':'', 'body':'Don\'t forget to follow!'})
+        # post[1].author = asker
+    if len(post) > 7:
+        post.insert (5, {'author':'', 'body':'If you hold the comment button, you will see your 4 most used emoji'})
+    if len(post) > 12:
+        post.insert (9, {'author':'', 'body':'The third name you see when you click Share then More secrelty has a crush on you'})
+        
     # if len(post) > 3:
-    #     post.insert(3, 'If you hold the comment button, you will see your 4 most used emoji')
+    #     post.insert(3, asker, 'If you hold the comment button, you will see your 4 most used emoji')
     #     post[1].author = asker
     #     commentBlankImage(post[3], 3, asker)
-
+    
     for j in range(len(post)):
-        
-
-        if post[j].author is None:
-            # set post author to [deleted] if no post author found
-            author = "[deleted]"
-        else:
-            try:
-                author = post[j].author.name
-            except:
+        try:
+            if post[j].author is None:
+                # set post author to [deleted] if no post author found
                 author = "[deleted]"
-        if j == 0:
-            # if start of post, print title caption with post name and info
-            print(post[j].title)
-            titleImage(post[j].title, author, "r/"+subreddit)
-        else:
-            if j == 2:
-                # makeshift way to add a call to action until a better method can be added - replaces comment content
-                text = 'Don\'t forget to follow!'
-            elif j == 5:
-                # makeshift way to add a call to action until a better method can be added - replaces comment content
-                text = 'Holding the comment button shows your 4 most used emojis'
             else:
-                # sets text to comment content
+                try:
+                    author = post[j].author.name
+                except:
+                    author = "[deleted]"
+        except AttributeError:
+            pass
+
+        if j == 0:
+            # if title post, print title caption with post name and info
+            # print(post[j].title)
+            titleImage(post[j].title, author, "r/"+subreddit)
+            soundifyAuthor(post[0].title, asker)
+        else:
+            # if not title post, create comment caption
+            try:
                 text = post[j].body
-            print(text)
+            except AttributeError:
+                text = post[j]['body']
+
+            # DEBUG
+            # print(text)
             sections = []
 
-            # control size of caption to keep it on the screen
+            # control size of caption to fit it on the screen vertically
             if len(text) > 500:
                 length = 0
                 nextSpace = 0
@@ -96,14 +115,23 @@ for i in range(5):
             
             # create captions and audio for thread comments
             for section in range(len(sections)):
-                commentImage(author, sections[section], j, section, asker)
-                soundifyComment(sections[section], j, section, asker)
+                if sections[section] != '':
+                    try:
+                        if post[j]['author'] == '':
+                            commentBlankImage(sections[section], j, asker)
+                            soundifyComment(sections[section], j, section, asker)
+                    except TypeError:
+                        commentImage(author, sections[section], j, section, asker)
+                        soundifyComment(sections[section], j, section, asker)
 
-    # create audio for title caption
-    soundifyAuthor(post[0].title, asker)
+    # # create audio for title caption
+    # soundifyAuthor(post[0].title, asker)
 
     if post[0].author is None:
         author = "[deleted]"
     else:
         author = post[0].author.name
     createVideo(author)
+
+# TikTok tags
+# #fyp #foryoupage #askreddit #reddit #askredditstories #redditreadings #reddittts
